refactor : common 패키지 정리
