+++ src/main/java/com/yapp/artie/domain/archive/domain/exhibit/Exhibit.java
@@ -4,7 +4,7 @@ package com.yapp.artie.domain.archive.domain.exhibit;
 import com.yapp.artie.domain.archive.domain.artwork.Artwork;
 import com.yapp.artie.domain.archive.domain.category.Category;
 import com.yapp.artie.domain.user.adapter.out.persistence.UserJpaEntity;
-import com.yapp.artie.global.common.BaseEntity;
+import com.yapp.artie.global.common.persistence.BaseEntity;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
