+++ src/main/java/com/yapp/artie/domain/archive/exception/CategoryAlreadyExistException.java
@@ -1,7 +1,7 @@
 package com.yapp.artie.domain.archive.exception;
 
-import com.yapp.artie.global.exception.common.BusinessException;
-import com.yapp.artie.global.exception.response.ErrorCode;
+import com.yapp.artie.global.common.exception.BusinessException;
+import com.yapp.artie.global.common.exception.ErrorCode;
 
 public class CategoryAlreadyExistException extends BusinessException {
