+++ src/main/java/com/yapp/artie/domain/question/domain/Question.java
@@ -1,7 +1,7 @@
 package com.yapp.artie.domain.question.domain;
 
 import com.yapp.artie.domain.user.adapter.out.persistence.UserJpaEntity;
-import com.yapp.artie.global.common.BaseEntity;
+import com.yapp.artie.global.common.persistence.BaseEntity;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
