+++ src/main/java/com/yapp/artie/domain/user/adapter/in/web/GetUserThumbnailController.java
@@ -3,7 +3,7 @@ package com.yapp.artie.domain.user.adapter.in.web;
 import com.yapp.artie.domain.user.adapter.out.persistence.UserJpaEntity;
 import com.yapp.artie.domain.user.application.port.in.GetUserThumbnailQuery;
 import com.yapp.artie.domain.user.application.port.in.GetUserThumbnailResponse;
-import com.yapp.artie.global.annotation.WebAdapter;
+import com.yapp.artie.global.common.annotation.WebAdapter;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
