+++ src/main/java/com/yapp/artie/domain/user/adapter/in/web/RegisterUserController.java
@@ -5,8 +5,8 @@ import com.yapp.artie.domain.user.application.port.in.RegisterUserResponse;
 import com.yapp.artie.domain.user.application.port.in.RegisterUserUseCase;
 import com.yapp.artie.domain.user.application.port.out.TokenParsingPort;
 import com.yapp.artie.domain.user.domain.ArtieToken;
-import com.yapp.artie.global.annotation.WebAdapter;
-import com.yapp.artie.global.exception.common.InvalidValueException;
+import com.yapp.artie.global.common.annotation.WebAdapter;
+import com.yapp.artie.global.common.exception.InvalidValueException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
