+++ src/main/java/com/yapp/artie/domain/user/application/service/GetUserService.java
@@ -3,7 +3,7 @@ package com.yapp.artie.domain.user.application.service;
 import com.yapp.artie.domain.user.application.port.in.GetUserQuery;
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.domain.User;
-import com.yapp.artie.global.annotation.UseCase;
+import com.yapp.artie.global.common.annotation.UseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
