+++ src/main/java/com/yapp/artie/domain/user/application/service/GetUserThumbnailService.java
@@ -6,7 +6,7 @@ import com.yapp.artie.domain.user.application.port.in.GetUserThumbnailQuery;
 import com.yapp.artie.domain.user.application.port.in.GetUserThumbnailResponse;
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.domain.User;
-import com.yapp.artie.global.annotation.UseCase;
+import com.yapp.artie.global.common.annotation.UseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
