+++ src/main/java/com/yapp/artie/domain/user/application/service/RenameUserService.java
@@ -4,7 +4,7 @@ import com.yapp.artie.domain.user.application.port.in.RenameUserUseCase;
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.application.port.out.UpdateUserStatePort;
 import com.yapp.artie.domain.user.domain.User;
-import com.yapp.artie.global.annotation.UseCase;
+import com.yapp.artie.global.common.annotation.UseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
