+++ src/main/java/com/yapp/artie/domain/user/application/service/UserWithdrawalService.java
@@ -7,7 +7,7 @@ import com.yapp.artie.domain.user.application.port.out.DeleteExternalUserPort;
 import com.yapp.artie.domain.user.application.port.out.DeleteUserPort;
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.domain.User;
-import com.yapp.artie.global.annotation.UseCase;
+import com.yapp.artie.global.common.annotation.UseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
