+++ src/main/java/com/yapp/artie/global/common/annotation/PersistenceAdapter.java
@@ -1,4 +1,4 @@
-package com.yapp.artie.global.annotation;
+package com.yapp.artie.global.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
