+++ src/main/java/com/yapp/artie/global/common/persistence/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.yapp.artie.global.common;
+package com.yapp.artie.global.common.persistence;
 
 import java.time.LocalDateTime;
 import javax.persistence.Column;
