+++ src/main/java/com/yapp/artie/global/filter/GlobalExceptionHandler.java
@@ -3,9 +3,9 @@ package com.yapp.artie.global.filter;
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.SdkClientException;
 import com.fasterxml.jackson.databind.exc.InvalidFormatException;
-import com.yapp.artie.global.exception.common.BusinessException;
-import com.yapp.artie.global.exception.response.ErrorCode;
-import com.yapp.artie.global.exception.response.ErrorResponse;
+import com.yapp.artie.global.common.exception.BusinessException;
+import com.yapp.artie.global.common.exception.ErrorCode;
+import com.yapp.artie.global.common.exception.ErrorResponse;
 import java.nio.file.AccessDeniedException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
